Split allocation into new file
