+++ pkg/controllers/node/ipam.go
@@ -82,85 +82,6 @@ func NewIPAMController(cfg config.NodeControllerConfig, c client.Interface, cs *
 
 }
 
-type allocation struct {
-	ip     string
-	handle string
-	attrs  map[string]string
-
-	timestamp     *time.Time
-	confirmedLeak bool
-}
-
-func (a *allocation) fields() log.Fields {
-	f := log.Fields{
-		"ip":     a.ip,
-		"handle": a.handle,
-		"node":   a.attrs[ipam.AttributeNode],
-	}
-
-	if a.isPodIP() {
-		ns := a.attrs[ipam.AttributeNamespace]
-		pod := a.attrs[ipam.AttributePod]
-		f["pod"] = fmt.Sprintf("%s/%s", ns, pod)
-	}
-
-	return f
-}
-
-func (a *allocation) node() string {
-	if node, ok := a.attrs[ipam.AttributeNode]; ok {
-		return node
-	}
-	return ""
-}
-
-func (a *allocation) markLeak(leakGracePeriod time.Duration) {
-	if a.timestamp == nil {
-		t := time.Now()
-		a.timestamp = &t
-		log.WithFields(a.fields()).Infof("Candidate IP leak")
-	}
-
-	if time.Since(*a.timestamp) > leakGracePeriod && !a.isConfirmedLeak() {
-		a.markConfirmedLeak()
-	}
-}
-
-func (a *allocation) markConfirmedLeak() {
-	if a.timestamp == nil {
-		log.WithFields(a.fields()).Warnf("Confirmed IP leak")
-	} else {
-		log.WithFields(a.fields()).Warnf("Confirmed IP leak after %s", time.Since(*a.timestamp))
-	}
-	a.confirmedLeak = true
-}
-
-func (a *allocation) markValid() {
-	if a.timestamp != nil {
-		log.WithFields(a.fields()).Infof("Confirmed valid IP after %s", time.Since(*a.timestamp))
-	}
-	a.confirmedLeak = false
-	a.timestamp = nil
-}
-
-func (a *allocation) isConfirmedLeak() bool {
-	return a.confirmedLeak
-}
-
-func (a *allocation) isPodIP() bool {
-	ns := a.attrs[ipam.AttributeNamespace]
-	pod := a.attrs[ipam.AttributePod]
-
-	return ns != "" && pod != ""
-}
-
-func (a *allocation) isTunnelAddress() bool {
-	ipip := a.attrs[ipam.AttributeType] == ipam.AttributeTypeIPIP
-	vxlan := a.attrs[ipam.AttributeType] == ipam.AttributeTypeVXLAN
-	wg := a.attrs[ipam.AttributeType] == ipam.AttributeTypeWireguard
-	return ipip || vxlan || wg
-}
-
 type ipamController struct {
 	rl        workqueue.RateLimiter
 	client    client.Interface
